+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased]
+## [1.1.0] - 2026-09-09
 
 ### Added
 - [#115] Chunked values: transparent handling of single values above the